Fix facsimile view zooming and panning
Previously:
* Panning always reseted when new panning started.
* Panning did not respect the zoom level.
* Zoom always reseted to center of the facsimile.
* Page scroll broke facsimile view when zoomed in.

Note: I couldn't make the mobile pinch zoom work in facsimile view. Everything else works now on mobile and desktop but in order to zoom you need to click the zoom buttons or go to the facsimile zoom view to pinch zoom there. This is a trade of in time usage and making the perfect.