page-title paddings and overflow on page-read
The title page now behaves like the intro page and has the same paddings.
The column area in page-read has now hidden scrollbars both in x- and y-direction when content doesn't spill. The x-scrollbar has been moved to the bottom of the page-read area instead of directly beneath the columns.