Change app fonts
The default fonts in the app have been changed. "Muli", which previously was used everywhere, has now been replaced by "Roboto" and "Source Serif 4". Roboto is the default font for the entire app, and is used in menus and md-texts and other static texts. Source Serif 4 is used in all TEI-texts.

The old Muli-font is still available, but has been updated and now uses the font name "Mulish".

The fonts in the app are now set using css custom properties (a.k.a. css variables). You can override the default fonts in custom.css by setting the following css custom properties. In this example the fonts are set back to the original Muli/Mulish font family, so copy this to custom.css in projects that wish to use the original font.

:root {
   --app-base-font-stack: "Mulish", sans-serif;
   --tei-font-stack: "Mulish", sans-serif;
}

There are a few other css custom properties defining font stacks in the app. You can override them as well in custom.css if you wish. They are:

--system-font-stack: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
--home-banner-font-stack: var(--app-base-font-stack);
--top-menu-font-stack: var(--app-base-font-stack);
--tei-blackletter-font-stack: "Roboto", var(--system-font-stack);
--tei-print-stamp-font-stack: Cambria, "Hoefler Text", Utopia, "Liberation Serif", "Nimbus Roman No9 L Regular", Times, "Times New Roman", serif;
--tei-editorial-note-font-stack: "Helvetica Neue", "Arial Nova", Helvetica, Arial, sans-serif;