Add support for linking to the same read text or commentary using only position id
In read texts and commentaries links to the same or other read texts or commentaries needed to have collectionId, publicationId, possible chapterId and positionId in @target in xml and @href in html. Now, if you only specify positionId in @target/@href, it is assumed the link refers to a position in the currently opened read text or commentary.

So if the current collectionId is 1, publicationId 2 and chapterId ch1, a link in the read text to position #pos1 in the same text can be written i xml as
<xref type="readingtext" target="#pos1">link</xref>
instead of
<xref type="readingtext" target="1 2 ch1 #pos1">link</xref>